+++ ormar/models/mixins/save_mixin.py
@@ -222,9 +222,7 @@ class SavePrepareMixin(RelationMixin, AliasMixin):
 
     @staticmethod
     async def _upsert_through_model(
-        instance: "Model",
-        previous_model: "Model",
-        relation_field: "ForeignKeyField",
+        instance: "Model", previous_model: "Model", relation_field: "ForeignKeyField",
     ) -> None:
         """
         Upsert through model for m2m relation.